+++ .github/labeler.yml
@@ -1,0 +1,15 @@
+frontend:
+  - "src/**/*"
+  - "src/assets/**/*"
+  - "src/components/**/*"
+  - "src/layouts/**/*"
+  - "src/views/**/*"
+
+public:
+  - "public/**/*"
+
+config:
+  - "vue.config.js"
+  - "*.config.js"
+  - "*.json"
+  - "*.yml"
