+++ .github/pull_request_template.md
@@ -1,0 +1,16 @@
+## 🔥 Summary
+무엇을 수정/추가했는지 한 줄 요약
+
+## 📂 Changes
+- 변경사항 1
+- 변경사항 2
+- 변경사항 3
+
+## 🧪 Test Notes
+- 확인한 테스트 내용
+- 영향 범위가 있다면 함께 작성
+
+## 📸 Screenshots (Optional)
+
+## 🔗 Related Issues
+Closes #
