+++ .github/workflows/add-to-project.yml
@@ -1,0 +1,15 @@
+name: Add PR to Frontend Project
+
+on:
+  pull_request:
+    types: [opened, reopened]
+
+jobs:
+  add-to-project:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Add PR to GitHub Project
+        uses: actions/add-to-project@v0.4.0
+        with:
+          project-url: https://github.com/orgs/KOSATM/projects/3
+          github-token: ${{ secrets.GITHUB_TOKEN }}
