+++ .github/workflows/auto-label-commit.yml
@@ -1,0 +1,12 @@
+name: Auto Label by Commit Message
+
+on:
+  pull_request:
+    types: [opened, synchronize, edited]
+
+jobs:
+  label-by-commit:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Apply labels based on commit messages
+        uses: bcoe/conventional-release-labels@v1
