+++ src/components/AppFooter.vue
@@ -6,11 +6,6 @@
         <span class="footer-title">Find My Seoul</span>
       </div>
 
-      <!-- <div class="footer-links">
-        <router-link to="/planner" class="footer-link">Planner</router-link>
-        <router-link to="/supporter" class="footer-link">Supporter</router-link>
-        <router-link to="/history" class="footer-link">History</router-link>
-      </div> -->
 
       <div class="footer-divider"></div>
