+++ src/views/supporter/SupporterMain.vue
@@ -174,7 +174,6 @@ import { ref, computed } from 'vue'
 import { useRouter } from 'vue-router'
 import PageHeader from '@/components/common/PageHeader.vue'
 import UploadSection from '@/components/travelgram/UploadSection.vue'
-import PlannerChat from '@/views/planner/edit/PlannerChat.vue'
 
 
 const router = useRouter()
