+++ src/views/supporter/image-ai/SupporterImageAISelectPlan.vue
@@ -114,13 +114,13 @@ const confirm = () => {
   if (!selectedOption.value) return
 
   if (selectedOption.value === 'add' || selectedOption.value === 'replace') {
-    // attempt to go to PlannerEdit; if route name isn't registered, fall back to SupporterImageAIChoicePlan
+    // attempt to go to EditPlan; if route name isn't registered, fall back to SupporterImageAIChoicePlan
     router.push({
       name: 'planedit',
       state: { item, mode: selectedOption.value },
       query: { mode: selectedOption.value, itemId: item?.id ?? '', itemName: item?.name ?? '' }
     }).catch(() => {
-      // fallback for environments where PlannerEdit route is not defined
+      // fallback for environments where EditPlan route is not defined
       router.push({
         name: 'SupporterImageAIChoicePlan',
         state: { item, mode: selectedOption.value },
